+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ CHANGELOG
 
  * fixed sqldump to display stdout on stderr allowing it to be piped correctly
  * added gondor env and env:set to view and set instance environment variables
+ * removed undocumented files.include functionality in favor of gondor env
 
 1.0.1
 =====